requirements.txt
Jinja2==3.1.2
livereload==2.6.3
python-dotenv==1.0.0
